api/requirements.txt
fastapi==0.109.2
pydantic==2.6.1
wikipedia==1.4.0
python-dotenv==1.0.1
requests==2.31.0
